add new module ValidCheck
Add a new module ValidCheck in src folder.
This module can check if the number of arguments is correct, if the type of subterms is correct and if the symbols are valid.
This new version can check the type of arguments that compute by other old functions.(for example:f(x,g(x,y)))
Correct some errors in ComputeTerm