generate random term
Here I changed the first second functions. These two functions now can only change the arguments of the whole term. It cannot work on the "function" arguments in this term. This operation will be done in the function changeFunction.
In the randomTerm, a term can be randomly operated with these four functions and get a new term.