add some test cases
Adding function "overlaps" to hiden the FunctionSymbol in the new signature, which is the same as the old one.
Correcting one problem in ValidCheck. Before if the ground term use the name of a function, it can't be detected.
Adding 15 test cases to check if all terms (not ground terms) generated with new signature are invalid.