change the function overlaps
this time it can check if a function is the same as the old one. It can check one function in the whole old signature.
The old "overlaps" can only check the new and old FunctionSymbol that hat the same order in signature. So some same FunctionSymbol cann't be detected.