shorten module list and use Maybe in "giveType"
in the list only giving the names of function that will be called by other modules.
In the ComputeTerm module I change the datatype, using Maybe Term. Nothing means, after searching the whole signature, it can't find a coressponging type for this term.
In the ValidCheck module I delete the giveType function. It calls ComputeTerm module and can directly use giveType in this module.