correct ValidCheck and add some more test cases
ValidCheck: correct "checkNumber". Now it can detect when a term is generated with function but has arguments. This error can be found. Changing the name of function "make" to "giveArgType". This name is more readable and corresponding to its real function. Lastly I change the order of these three called functions in "inValid". The aim is, when the term contains some symbols that don't contain in signature, it must be invalid, the next two functions don't need to work. This can save time.
ComputeTerm: change some hlint problems
Signatures: add signature6. This signature contains some variables with same type.
Spec: add corresponding test cases to test if signatures like signature6 can be correctly deal with.