change checkType in ValidCheck
Previously I think if the type of this term don't exis in this sigature, then the result is false. But if the type don't exist that means the symbol of this term also don't exist in this signature. This can be detected with CheckSymbol. So here I deleted the first argument of checkType.