add a new module GetSignatureInfo
Adding a new module. Functions, which get or deal with information from signatures are now in this module.
At the same time, I changed the calling relationship in each Module.
In the test, I add a new function. This function can choose one function symbol from the new generated signature. Then this one will be used to compute invalid terms. This way, the amount of terms is smaller. The test time will be shorter.