delete randomSigTotal' function in ArbitrarySig
I think this function has some problems. With with function a signature may only contain functions.
That means no terms can be generated with this siganture.
I think the result of randomSigTotal and randomSigTotal' is fast same. But randomSigTotal save the original constants in a signature. As a result, I think this function is more useful.
According to this change, I also delete some test cases.