oneValidTerm now with size
Now this function can generate valid terms with given number of nodes. But I think this is not perfect. Because now it can only check the first level of its subterms. How to control the subterm of subterms?
Here is also a problem. If I only write n<=b (b is the high limit), it can work well. But if I only write n>=a (a is the low limit), when a is larger than 1, the results are all Nothing. Now I don't know why this happens.
So here I also write n<=b, it can meet the requirement.