OneTerm with new split and check solution
I found if the interval is for example, between 1 and 4, no subdivision can be found (with division function). But this is not correct, the result is possible not Nothing. Constants and terms with size 3 can be generated.
So here I add a new condition, if the lower bound is 1 and the selected function symbol has no arguments, directly generating a constant. Or the lower bound will be changed to the length of arguments list. In this way, we may find a possible division.
In DataType, when I added {-# LANGUAGE OverloadedRecordFields #-}, an error said, this is an unsupported extension. I use {-# LANGUAGE DuplicateRecordFields#-}, it cannot work. So where should I input this extension, in cabal?