oneValidTerm with Mode, delete outdated fiels
Outdated files in app are deleted.
In OneTerm oneValidTerm is now with Mode. But the third Mode "ONCE" doesn't work well. Now it can generate a given symbol in term exactly once, but it needs a check in oneValidTerm. Without this check, it will generate a term, which contains at most once this symbol.