Mode with new constructor and OneTerm with successful method
Mode is now in original version.
In OneTerm using newModes to generate a list of Mode. If the mode is ONCE, it selects one node and change its mode to ONCE. Other nodes are NO.
Then using zip3 and map to make it work in the almost same way as before.