delete two useless modules, move Mode to OneTerm
ArbitraryTerm and ComputeAllTerm these two modules are at present useless. So I deleted them.
Data constructor is moved to OneTerm. isConstant is used to check if this function symbol is constant. isOnce is used to check if this Mode is ONCE s.
Adding some test cases to test this module.