change CertainSignature, wrongSymbol' and some test cases
In CertainSignature, when not enough correct terms can be generated, there is a tip. The number of incorrect terms is garanteed.
wrongSymbol' produces a new signature, whose input and ouput types are different from the old one.
The test cases used to test invalidTerms have been corrected. With this judgement, an empty list is also considered.